HWKALERTS-119 Move Bus messages to public API
