+++ hawkular-alerts-bus-api/src/main/java/org/hawkular/alerts/bus/api/AlertDataMessage.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates
+ * Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.alerts.bus.messages;
+package org.hawkular.alerts.bus.api;
 
 
 import java.util.Arrays;
@@ -24,6 +24,7 @@ import org.hawkular.alerts.api.model.data.Data;
 import org.hawkular.bus.common.AbstractMessage;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+
 /**
  * A bus message used to receive data for the alerts subsystem.
  * One message can store a collection of {@link org.hawkular.alerts.bus.messages.AlertData}.
