+++ hawkular-alerts-bus-api/src/main/java/org/hawkular/alerts/bus/api/AvailDataMessage.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates
+ * Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.alerts.bus.messages;
+package org.hawkular.alerts.bus.api;
 
 import java.util.List;
